requirements.txt
fastapi
uvicorn
playwright
python-multipart
psycopg2-binary
